INTYG-6719: Refactored AccessService implementations. Still some refactoring on Draft and LockedDraft left.
